requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
APScheduler==3.10.4
Werkzeug==3.0.1
WTForms==3.1.1
bcrypt==4.1.2
requests==2.31.0
Pillow==10.1.0
python-barcode==0.15.1
pyzbar==0.1.9
torchvision==0.16.2
torch==2.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
playwright==1.40.0
pytest-playwright==0.4.3
black==23.11.0
isort==5.12.0
flake8==6.1.0
pre-commit==3.5.0